Fix `PolyGrid` bug by requiring 3 layers, plus more tweaks to raycasts.
* `PolyGrid` bug tracking child `Poly`s to update, caused by `Poly`s
requiring a point in a neighbouring `Poly` to be polygonised that is not
tracked with 2 layers (far corner of cube) -- could not fix, so gave up
and now require a `Surface` to have (at least) 3 layers in order to
polygonise it.  This simplifies the polygonisation algorithm hugely, but
has a big tradeoff in efficiency of narrow band updates.
* Fixed `box_intersect` AABB intersection to skip over candidate 
intersects that are not in bounding box and try next one.
* Surface updates now return unused quantity (because of clamping or
whatever).
* Raycasts now interpolate, rather than round, sample points.
* Raycast now moves at half speed to ensure all candidate child grids
and leaf nodes are considered.
* Added lookup grid `reset_all` method and refactored `reset` as first
victim of GitHub issue #1 to remove virtual methods where possible.
* Optimisation to polygonisation to return early when creating simplices
if there are none to create.
* Optimise `walk_band` to be templated and use static member, so lookup
grid is re-used.