[#6] `Surface` deactivate with correct background value + fix raycast
* Fixed: `Surface::layer_remove` now passes background value to use when
removing a point, in case the host child is to be `deactivate`d, so that
"inside" deactivated child grids have an "inside" background value, and
"outside" deactivated has an "outside" background.
* Fixed: `Surface::ray_check_add_child` will now consider children for
ray tests if they contain layers +/-1 as well as 0, for the corner case
of rays that hit precisely the corner of a child with zero layer points
just over the border.
* Fixed: `TrackedGridBase` constructor to not call base constructor as
well as own `init`, since lookup grid is not defined in subsequent calls
in base construction.
* Fixed: Renamed `self` macro to `nself` (for non-const self), due to
clash with Eigen.
* Added: `LazySingleTrackedPartitionedGrid` now has overrides for
`remove` and `remove_child` that takes a background value to use when a
child grid is to be deactivated.
* Added: Optimisation tweak to `StaticGridBase::interp` to modify a
single `std::vector` in-line, rather than creating/destroying
temporarys. Did this after OProfile profile run using "UrFelt" testbed
found `interp` as the taking the most CPU cycles.
* Added: `ceil` utility to round up float vectors to int vectors.
* Added: Extra debugging flags to CMake compiler options.
* Added: Some tidying, docstrings and better error reporting when
`FELT_EXCEPTIONS` defined.
* Added/Fixed: Tests updated and passing.