[#11] remove temporarily added child partitions during raycast
* In order to allow raycasting to continue through child partitions that
overhang the border of the grid (because they are not an exact factor of
the grid size), we allowed child partitions to be added as candidates
even if the ray hit point was outside the grid (as long as the ray is
travelling in the right direction).  
* But if the hit point is outside the grid then `pos_child()` returns an
undefined value (in particular, when negatively outside the grid), which
can resolve to a valid child partition, but with a ray hit point outside
of it.
* If this invalid hit point is closer to the ray origin then that
`ChildHit` is preferred over a valid one when doing the
sort-unique-erase.
* Then when casting through the child, `ray()` returns immediately
thinking the ray has already passed outside the child.
* So reintroduced a check for if `pos_intersect` is `inside` the grid in
`ray_check_add_child`, but still `return true` so that subsequent child
partitions along the ray's path can still be considered.