Fix `LookupGridBase::remove` to no-op if pos not tracked
* This caused long, hard to debug problems with `PolyGrid::notify`
gradually failing to add `Poly` children to the tracking list to be
repolyed, because when it `remove`s `Poly` children from tracking it
doesn't check if they are tracked or not.  So the untracked pos was left
unmodified, but the tracking list still popped, removing an unrelated
`Poly` from tracking(!).
* So made `remove` behave similar to `add`, in that if a point is not
tracked, `remove` is a no-op.
* Also added regression test for this case.